requirements.txt
antlr4-python3-runtime>=4.9.3
cycler>=0.11.0
dataclasses>=0.8
decorator>=4.4.2
dill>=0.3.4
easydict>=1.10
hydra-core~=1.3.2
imageio>=2.15.0
importlib-resources>=5.4.0
inplace-abn>=1.1.0
joblib>=1.1.1
kiwisolver>=1.3.1
matplotlib>=3.3.4
multiprocess>=0.70.12.2
networkx>=2.5.1
numpy>=1.19.5
omegaconf>=2.3.0
p-tqdm>=1.4.0
packaging>=21.3
pathos>=0.2.8
Pillow>=8.4.0
pox>=0.3.0
ppft>=1.6.6.4
pyparsing>=3.0.9
python-dateutil>=2.8.2
pytz>=2023.2
PyWavelets>=1.1.1
PyYAML>=6.0
scikit-image>=0.17.2
scikit-learn>=0.24.2
scipy>=1.5.4
seaborn>=0.11.2
six>=1.16.0
sklearn>=0.0
threadpoolctl>=3.1.0
tifffile>=2020.9.3
torch>=1.9.0+cu111
torchaudio>=0.9.0
torchvision>=0.10.0+cu111
tqdm>=4.64.1
typing_extensions>=4.1.1
zipp>=3.6.0
